+++ include/aquarius/context/invoke.hpp
@@ -1,29 +1,65 @@
 #pragma once
-#include <aquarius/context/generator.hpp>
+#include <aquarius/context/manager.hpp>
+#include <aquarius/context/impl/session.hpp>
 
 namespace aquarius
 {
-	class basic_context;
-}
-
-namespace aquarius
-{
-	template <typename _Context>
-	struct ctx_regist
+	struct invoke_session_helper
 	{
-		ctx_regist(std::size_t key)
+		static bool push(std::shared_ptr<basic_session> session_ptr)
 		{
-			context_generator::instance().regist(key, []()
-										  { return std::dynamic_pointer_cast<basic_context>(std::make_shared<_Context>()); });
+			return session_manager::instance().push(session_ptr);
 		}
-	};
 
-	struct invoke_context_helper
-	{
-		template <typename... _Args>
-		static auto invoke(std::size_t key, _Args&... args)
+		static std::shared_ptr<basic_session> find(std::size_t uid)
 		{
-			return context_generator::instance().invoke(key, args...);
+			return session_manager::instance().invoke(uid);
+		}
+
+		static void erase(std::size_t uid)
+		{
+			return session_manager::instance().erase(uid);
+		}
+
+		static std::size_t size()
+		{
+			return session_manager::instance().size();
+		}
+
+		template <typename _Response>
+		static void broadcast(_Response&& resp)
+		{
+			error_code ec{};
+
+			flex_buffer_t fs{};
+			resp.to_binary(fs, ec);
+
+			if (ec)
+			{
+				XLOG_ERROR() << "[Message] " << resp.uuid() << "to binary occur error.";
+
+				return;
+			}
+
+			session_manager::instance().broadcast(std::move(fs));
+		}
+
+		template <typename _Response, typename _Pre>
+		static void broadcast_if(_Response&& resp, _Pre&& func)
+		{
+			error_code ec{};
+
+			flex_buffer_t fs{};
+			resp.to_binary(fs, ec);
+
+			if (ec)
+			{
+				XLOG_ERROR() << "[Message] " << resp.uuid() << "to binary occur error.";
+
+				return;
+			}
+
+			session_manager::instance().broadcast_if(std::move(fs), std::forward<_Pre>(func));
 		}
 	};
 } // namespace aquarius
