+++ include/aquarius/context/session.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <any>
-#include <aquarius/connect/impl/session.hpp>
+#include <aquarius/context/impl/session.hpp>
 #include <deque>
 #include <memory>
 #include <string>
