+++ test/unit_test.vcxproj
@@ -165,18 +165,16 @@
     <ClInclude Include="..\include\aquarius\client\async.hpp" />
     <ClInclude Include="..\include\aquarius\client\callback.hpp" />
     <ClInclude Include="..\include\aquarius\connect.hpp" />
-    <ClInclude Include="..\include\aquarius\connect\impl\session.hpp" />
-    <ClInclude Include="..\include\aquarius\connect\invoke.hpp" />
-    <ClInclude Include="..\include\aquarius\connect\manager.hpp" />
-    <ClInclude Include="..\include\aquarius\connect\session.hpp" />
-    <ClInclude Include="..\include\aquarius\connect\impl\socket_adapter.hpp" />
+    <ClInclude Include="..\include\aquarius\connect\socket_adapter.hpp" />
     <ClInclude Include="..\include\aquarius\context.hpp" />
     <ClInclude Include="..\include\aquarius\context\content.hpp" />
     <ClInclude Include="..\include\aquarius\context\context.hpp" />
     <ClInclude Include="..\include\aquarius\context\generator.hpp" />
     <ClInclude Include="..\include\aquarius\context\impl\context.hpp" />
-    <ClInclude Include="..\include\aquarius\context\impl\visitor.hpp" />
+    <ClInclude Include="..\include\aquarius\context\impl\session.hpp" />
     <ClInclude Include="..\include\aquarius\context\invoke.hpp" />
+    <ClInclude Include="..\include\aquarius\context\manager.hpp" />
+    <ClInclude Include="..\include\aquarius\context\session.hpp" />
     <ClInclude Include="..\include\aquarius\core\apply.hpp" />
     <ClInclude Include="..\include\aquarius\core\error_code.hpp" />
     <ClInclude Include="..\include\aquarius\core\generator.hpp" />
@@ -187,6 +185,7 @@
     <ClInclude Include="..\include\aquarius\core\impl\time_traits.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\type_traits.hpp" />
     <ClInclude Include="..\include\aquarius\core\manager.hpp" />
+    <ClInclude Include="..\include\aquarius\core\visitor.hpp" />
     <ClInclude Include="..\include\aquarius\message\generator.hpp" />
     <ClInclude Include="..\include\aquarius\message\header.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\custom.hpp" />
